+++ Domain/RDD.Domain/Patchers/PatcherProvider.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using NExtends.Primitives.Types;
 using Rdd.Domain.Exceptions;
+using Rdd.Domain.Helpers.Reflection;
 using Rdd.Domain.Json;
 using System;
 using System.Collections;
@@ -10,10 +11,12 @@ namespace Rdd.Domain.Patchers
     public class PatcherProvider : IPatcherProvider
     {
         protected IServiceProvider Services { get; set; }
+        protected IReflectionHelper IReflectionHelper { get; set; }
 
-        public PatcherProvider(IServiceProvider services)
+        public PatcherProvider(IServiceProvider services, IReflectionHelper reflectionHelper)
         {
             Services = services ?? throw new ArgumentNullException(nameof(services));
+            IReflectionHelper = reflectionHelper ?? throw new ArgumentNullException(nameof(reflectionHelper));
         }
 
         public virtual IPatcher GetPatcher(Type expectedType, IJsonElement json)
@@ -33,7 +36,7 @@ namespace Rdd.Domain.Patchers
                 return Services.GetService<DictionaryPatcher>();
             }
 
-            if (expectedType == typeof(string) || expectedType.IsValueType || expectedType.GetNullableType().IsValueType)
+            if (IReflectionHelper.IsPseudoValue(expectedType) || expectedType.GetNullableType().IsValueType)
             {
                 return Services.GetService<ValuePatcher>();
             }